Augment data with information from a model_stack object

Augment accepts a model object and a dataset and adds information about each observation in the dataset.

# S3 method for model_stack
augment(x, data = x[["train"]], ...)

Arguments

x
A model_stack object

data
A data.frame-like object containing all of the necessary predictors that the model stack was trained with.

...
Additional arguments. Currently ignored.

Value

A tibble::tibble() with additional information on each observation appended.

Example Data

This package provides some resampling objects and datasets for use in examples and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal 7ish days if they detect potential predator threat. Researchers wanted to determine how, and when, these tree frog embryos were able to detect stimulus from their environment. To do so, they subjected the embryos at varying developmental stages to "predator stimulus" by jiggling the embryos with a blunt probe. Beforehand, though some of the embryos were treated with gentamicin, a compound that knocks out their lateral line (a sensory organ.) Researcher Julie Jung and her crew found that these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily a representative or unbiased subset of the complete dataset, and is only for demonstrative purposes.

reg_folds and class_folds are rset cross-fold validation objects from rsample, splitting the training data into for the regression and classification model objects, respectively. tree_frogs_reg_test and tree_frogs_class_test are the analogous testing sets.

reg_res_lr, reg_res_svm, and reg_res_sp contain regression tuning results for a linear regression, support vector machine, and spline model, respectively, fitting latency (i.e. how long the embryos took to hatch in response to the jiggle) in the tree_frogs data, using most all of the other variables as predictors. Note that the data underlying these models is filtered to include data only from embryos that hatched in response to the stimulus.

class_res_rf and class_res_nn contain multiclass classification tuning results for a random forest and neural network classification model, respectively, fitting reflex (a measure of ear function) in the data using most all of the other variables as predictors.

log_res_rf and log_res_nn, contain binary classification tuning results for a random forest and neural network classification model, respectively, fitting hatched (whether or not the embryos hatched in response to the stimulus) using most all of the other variables as predictors.

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

See also

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

# to use the `tidy` method, the appropriate glance
# methods must be loaded for the member model tidiers.
# in this example, they come from broom!
library(broom)

# build a regression model stack
st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()
  
# check the model stack out!
tidy(st)

#> # A tibble: 4 x 13
#>   member r.squared adj.r.squared sigma statistic   p.value    df logLik   AIC
#>   <chr>      <dbl>         <dbl> <dbl>     <dbl>     <dbl> <dbl>  <dbl> <dbl>
#> 1 reg_r…     0.384         0.375  58.7      43.8  1.45e-41     6 -2352. 4720.
#> 2 reg_r…     0.395         0.385  58.2      39.3  1.92e-42     7 -2348. 4714.
#> 3 reg_r…     0.422         0.395  57.7      15.7  7.00e-38    19 -2338. 4719.
#> 4 ensem…    NA            NA      NA        NA   NA           NA    NA    NA 
#> # … with 4 more variables: BIC <dbl>, deviance <dbl>, df.residual <int>,
#> #   nobs <int>

glance(st)

#> # A tibble: 1 x 12
#>   r.squared adj.r.squared sigma statistic p.value    df logLik   AIC   BIC
#>       <dbl>         <dbl> <dbl>     <dbl>   <dbl> <int>  <dbl> <dbl> <dbl>
#> 1        NA            NA    NA        NA      NA    NA     NA    NA    NA
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>

augment(st)

#> # A tibble: 429 x 6
#>    treatment  reflex    age t_o_d     latency .fitted$.pred
#>    <chr>      <fct>   <dbl> <chr>       <dbl>         <dbl>
#>  1 control    low    361180 night         360         194. 
#>  2 control    full   401595 afternoon     106         114. 
#>  3 control    mid    357810 night         180         123. 
#>  4 control    full   397440 afternoon      60         112. 
#>  5 gentamicin full   463230 morning        39          84.7
#>  6 control    full   393900 afternoon     214         120. 
#>  7 control    full   469065 morning        50          35.8
#>  8 control    full   400240 afternoon     224         111. 
#>  9 control    full   466160 morning        63          38.2
#> 10 control    full   467950 morning        33          39.4
#> # … with 419 more rows

# }