Evaluate a data stack

Evaluates a stack by performing regularization on the out-of-sample predictions to determine coefficients for the combining of predictions from ensemble members.

blend_predictions(data_stack, penalty = 10^(-6:-1), verbose = FALSE, ...)

Arguments

data_stack
A data_stack object

penalty
A numeric vector of proposed penalty values used in member weighting. Higher penalties will generally result in fewer members being included in the resulting model stack, and vice versa. This argument will be tuned on unless a single penalty value is given.

verbose
A logical for logging results as they are generated. Despite this argument, warnings and errors are always shown.

...
Additional arguments. Currently ignored.

Value

A model_stack object—while model_stacks largely contain the same elements as data_stacks, the primary data objects shift from the assessment set predictions to the member models.

Example Data

These examples make use of data and resampling objects exported with the package. All of them are derived from test cases making use of the penguins dataset from Allison Horst's palmerpenguins package.

The penguins_train and penguins_test objects are subsets of the penguins data for using in training and testing, respectively.

Objects containing the substring _res_ are tune_results objects for model specifications. The reg prefix indicates that the model definition is for use in regression, class indicates multinomial classification, and log (as in logistic) indicates binomial classification. The suffix indicates the model definition type. All of these tune_results objects are fitted on a 5-fold cross validation of the penguins_train data.

For the regression setting, these tune_results objects reflect models specified to fit body_mass_g using all of the other variables as predictors. These objects include:

reg_res_lr: Fitted resamples for a linear regression model

reg_res_sp: Tuning results for a splines model

reg_res_svm: Tuning results for a support vector machine model

In the multinomial classification setting, the relevant objects reflect models specified to fit year (as a factor) using all of the other variables as predictors. These objects include:

class_res_nn: Fitted resamples for a neural network model

class_res_rf: Tuning results for a random forest model

In the binomial classification setting, the relevant objects reflect models specified to fit sex using all of the other variables as predictors. These objects include:

log_res_nn: Fitted resamples for a neural network model

log_res_rf: Tuning results for a random forest model

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

See also

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the objects used in these examples!

# put together a data stack
reg_st <- 
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_svm) %>%
  add_candidates(reg_res_sp)
  
reg_st

#> # A data stack with 3 model definitions and 16 candidate members:
#> #   reg_res_lr: 1 sub-model
#> #   reg_res_svm: 5 sub-models
#> #   reg_res_sp: 10 sub-models
#> # Outcome: body_mass_g (integer)

# evaluate the data stack
reg_st %>%
  blend_predictions()

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 16 possible candidates, the ensemble contains 4 members.

#> The highest weighted members are:

#> # A tibble: 4 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_lr1  linear_reg  0.374
#> 2 reg_res_sp06 linear_reg  0.248
#> 3 reg_res_svm3 svm_rbf     0.235
#> 4 reg_res_svm5 svm_rbf     0.191

#> 
#> Members have not yet been fitted with `fit_members()`.

# include fewer models by proposing
# higher penalties
reg_st %>% blend_predictions(penalty = c(.5, 1))

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 16 possible candidates, the ensemble contains 4 members.

#> The highest weighted members are:

#> # A tibble: 4 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_lr1  linear_reg  0.374
#> 2 reg_res_sp06 linear_reg  0.248
#> 3 reg_res_svm3 svm_rbf     0.235
#> 4 reg_res_svm5 svm_rbf     0.191

#> 
#> Members have not yet been fitted with `fit_members()`.

  
# do the same with multinomial classification models
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions()

#> ! Fold1: model 1/1: from glmnet Fortran code (error code -58); Convergence for 58...

#> ! Fold2: model 1/1: from glmnet Fortran code (error code -65); Convergence for 65...

#> ! Fold3: model 1/1: from glmnet Fortran code (error code -68); Convergence for 68...

#> ! Fold4: model 1/1: from glmnet Fortran code (error code -62); Convergence for 62...

#> ! Fold5: model 1/1: from glmnet Fortran code (error code -63); Convergence for 63...

  
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 22 possible candidates, the ensemble contains 13 members.

#> Across the 3 classes, there are an average of 4.33 members per class.

#> The highest weighted member classes are:

#> # A tibble: 10 x 4
#>    member                    type        weight class
#>    <chr>                     <chr>        <dbl> <chr>
#>  1 .pred_2009_class_res_rf01 rand_forest  7.19  2007 
#>  2 .pred_2009_class_res_rf01 rand_forest  5.35  2008 
#>  3 .pred_2009_class_res_nn1  mlp          3.31  2008 
#>  4 .pred_2009_class_res_nn1  mlp          3.06  2009 
#>  5 .pred_2009_class_res_rf06 rand_forest  2.46  2009 
#>  6 .pred_2009_class_res_rf04 rand_forest  2.07  2009 
#>  7 .pred_2008_class_res_rf02 rand_forest  1.12  2008 
#>  8 .pred_2008_class_res_nn1  mlp          0.899 2008 
#>  9 .pred_2008_class_res_rf05 rand_forest  0.836 2009 
#> 10 .pred_2008_class_res_rf07 rand_forest  0.727 2008 

#> 
#> Members have not yet been fitted with `fit_members()`.

# ...or binomial classification models
log_st <-
  stacks() %>%
  add_candidates(log_res_nn) %>%
  add_candidates(log_res_rf) %>%
  blend_predictions()
  
log_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 11 possible candidates, the ensemble contains 3 members.

#> The highest weighted member classes are:

#> # A tibble: 3 x 3
#>   member                  type        weight
#>   <chr>                   <chr>        <dbl>
#> 1 .pred_male_log_res_nn1  mlp           5.50
#> 2 .pred_male_log_res_rf10 rand_forest   3.85
#> 3 .pred_male_log_res_rf06 rand_forest   1.30

#> 
#> Members have not yet been fitted with `fit_members()`.

# }