Fit model stack members

Fit model stack members

fit_members(model_stack, verbose = FALSE, ...)

Arguments

model_stack
A model_stack object outputted by blend_predictions() or fit_members()

verbose
A logical for logging results as they are generated. Despite this argument, warnings and errors are always shown.

...
Additional arguments. Currently ignored.

Value

A model_stack object with a subclass inherited from the chosen *_stack method---this fitted model contains the necessary components to predict on new data.

Example Data

This package provides some resampling objects and datasets for use in examples and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal 7ish days if they detect potential predator threat. Researchers wanted to determine how, and when, these tree frog embryos were able to detect stimulus from their environment. To do so, they subjected the embryos at varying developmental stages to "predator stimulus" by jiggling the embryos with a blunt probe. Beforehand, though some of the embryos were treated with gentamicin, a compound that knocks out their lateral line (a sensory organ.) Researcher Julie Jung and her crew found that these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily a representative or unbiased subset of the complete dataset, and is only for demonstrative purposes.

reg_folds and class_folds are rset cross-fold validation objects from rsample, splitting the training data into for the regression and classification model objects, respectively. tree_frogs_reg_test and tree_frogs_class_test are the analogous testing sets.

reg_res_lr, reg_res_svm, and reg_res_sp contain regression tuning results for a linear regression, support vector machine, and spline model, respectively, fitting latency (i.e. how long the embryos took to hatch in response to the jiggle) in the tree_frogs data, using most all of the other variables as predictors. Note that the data underlying these models is filtered to include data only from embryos that hatched in response to the stimulus.

class_res_rf and class_res_nn contain multiclass classification tuning results for a random forest and neural network classification model, respectively, fitting reflex (a measure of ear function) in the data using most all of the other variables as predictors.

log_res_rf and log_res_nn, contain binary classification tuning results for a random forest and neural network classification model, respectively, fitting hatched (whether or not the embryos hatched in response to the stimulus) using most all of the other variables as predictors.

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

See also

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the objects used in these examples!

# put together a data stack
reg_st <- 
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_svm) %>%
  add_candidates(reg_res_sp)
  
reg_st

#> # A data stack with 3 model definitions and 16 candidate members:
#> #   reg_res_lr: 1 sub-model
#> #   reg_res_svm: 5 sub-models
#> #   reg_res_sp: 10 sub-models
#> # Outcome: latency (numeric)

# evaluate the data stack and fit the member models
reg_st %>%
  blend_predictions() %>%
  fit_members()

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 16 possible candidates, the ensemble contains 5 members.

#> The highest weighted members are:

#> # A tibble: 5 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_svm4 svm_rbf    1.58  
#> 2 reg_res_svm3 svm_rbf    0.936 
#> 3 reg_res_sp06 linear_reg 0.216 
#> 4 reg_res_sp10 linear_reg 0.124 
#> 5 reg_res_svm5 svm_rbf    0.0450

  
reg_st

#> # A data stack with 3 model definitions and 16 candidate members:
#> #   reg_res_lr: 1 sub-model
#> #   reg_res_svm: 5 sub-models
#> #   reg_res_sp: 10 sub-models
#> # Outcome: latency (numeric)

  
# do the same with multinomial classification models
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()

#> ! Fold2: model 1/1: from glmnet Fortran code (error code -87); Convergence for 87...

#> ! Fold5: model 1/1: from glmnet Fortran code (error code -84); Convergence for 84...

  
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 22 possible candidates, the ensemble contains 5 members.

#> Across the 3 classes, there are an average of 2.5 members per class.

#> The highest weighted member classes are:

#> # A tibble: 5 x 4
#>   member                    type         weight class
#>   <chr>                     <chr>         <dbl> <chr>
#> 1 .pred_full_class_res_nn1  mlp         19.4    full 
#> 2 .pred_mid_class_res_rf06  rand_forest  3.47   mid  
#> 3 .pred_full_class_res_rf05 rand_forest  1.65   full 
#> 4 .pred_full_class_res_rf03 rand_forest  0.315  full 
#> 5 .pred_full_class_res_rf02 rand_forest  0.0477 full 

  
# ...or binomial classification models
log_st <-
  stacks() %>%
  add_candidates(log_res_nn) %>%
  add_candidates(log_res_rf) %>%
  blend_predictions() %>%
  fit_members()
  
log_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 11 possible candidates, the ensemble contains 2 members.

#> The highest weighted member classes are:

#> # A tibble: 2 x 3
#>   member                 type        weight
#>   <chr>                  <chr>        <dbl>
#> 1 .pred_yes_log_res_nn1  mlp           7.93
#> 2 .pred_yes_log_res_rf05 rand_forest   2.83

# }