Model predictions

Apply a model stack to create different types of predictions.

# S3 method for model_stack
predict(object, new_data, type = NULL, members = FALSE, opts = list(), ...)

Arguments

object
A model stack with fitted members.

new_data
A rectangular data object, such as a data frame.

type
Format of returned predicted values—one of "numeric", "class", or "prob". When NULL, predict() will choose an appropriate value based on the model's mode.

members
Logical. Whether or not to additionally return the predictions for each of the ensemble members.

opts
A list of optional arguments to the underlying predict function passed on to parsnip::predict.model_fit for each member.

...
Additional arguments. Currently ignored.

Example Data

This package provides some resampling objects and datasets for use in examples and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal 7ish days if they detect potential predator threat. Researchers wanted to determine how, and when, these tree frog embryos were able to detect stimulus from their environment. To do so, they subjected the embryos at varying developmental stages to "predator stimulus" by jiggling the embryos with a blunt probe. Beforehand, though some of the embryos were treated with gentamicin, a compound that knocks out their lateral line (a sensory organ.) Researcher Julie Jung and her crew found that these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily a representative or unbiased subset of the complete dataset, and is only for demonstrative purposes.

reg_folds and class_folds are rset cross-fold validation objects from rsample, splitting the training data into for the regression and classification model objects, respectively. tree_frogs_reg_test and tree_frogs_class_test are the analogous testing sets.

reg_res_lr, reg_res_svm, and reg_res_sp contain regression tuning results for a linear regression, support vector machine, and spline model, respectively, fitting latency (i.e. how long the embryos took to hatch in response to the jiggle) in the tree_frogs data, using most all of the other variables as predictors. Note that the data underlying these models is filtered to include data only from embryos that hatched in response to the stimulus.

class_res_rf and class_res_nn contain multiclass classification tuning results for a random forest and neural network classification model, respectively, fitting reflex (a measure of ear function) in the data using most all of the other variables as predictors.

log_res_rf and log_res_nn, contain binary classification tuning results for a random forest and neural network classification model, respectively, fitting hatched (whether or not the embryos hatched in response to the stimulus) using most all of the other variables as predictors.

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

data(tree_frogs_reg_test)
data(tree_frogs_class_test)

# build and fit a regression model stack
reg_st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()

reg_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 11 possible candidates, the ensemble contains 3 members.

#> The highest weighted members are:

#> # A tibble: 3 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_sp10 linear_reg  0.379
#> 2 reg_res_lr1  linear_reg  0.335
#> 3 reg_res_sp06 linear_reg  0.209

# predict on the tree frogs testing data
predict(reg_st, tree_frogs_reg_test)

#> # A tibble: 143 x 1
#>    .pred
#>    <dbl>
#>  1  40.0
#>  2 115. 
#>  3  88.9
#>  4  34.3
#>  5  76.5
#>  6  88.2
#>  7 123. 
#>  8  80.7
#>  9  37.1
#> 10  77.2
#> # … with 133 more rows

# include the predictions from the members
predict(reg_st, tree_frogs_reg_test, members = TRUE)

#> # A tibble: 143 x 4
#>    .pred reg_res_lr1 reg_res_sp10 reg_res_sp06
#>    <dbl>       <dbl>        <dbl>        <dbl>
#>  1  40.0        38.0         34.3         37.5
#>  2 115.        124.         117.         111. 
#>  3  88.9        84.5         92.4         92.0
#>  4  34.3        35.3         28.0         25.5
#>  5  76.5        79.0         77.0         69.4
#>  6  88.2        83.7         92.2         89.9
#>  7 123.        118.         139.         116. 
#>  8  80.7        80.0         79.2         83.8
#>  9  37.1        36.5         30.7         32.6
#> 10  77.2        79.4         78.0         70.1
#> # … with 133 more rows

# build and fit a classification model stack
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()

#> ! Fold2: model 1/1: from glmnet Fortran code (error code -87); Convergence for 87...

#> ! Fold5: model 1/1: from glmnet Fortran code (error code -84); Convergence for 84...

 
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 22 possible candidates, the ensemble contains 5 members.

#> Across the 3 classes, there are an average of 2.5 members per class.

#> The highest weighted member classes are:

#> # A tibble: 5 x 4
#>   member                    type         weight class
#>   <chr>                     <chr>         <dbl> <chr>
#> 1 .pred_full_class_res_nn1  mlp         19.4    full 
#> 2 .pred_mid_class_res_rf06  rand_forest  3.47   mid  
#> 3 .pred_full_class_res_rf05 rand_forest  1.65   full 
#> 4 .pred_full_class_res_rf03 rand_forest  0.315  full 
#> 5 .pred_full_class_res_rf02 rand_forest  0.0477 full 

# predict reflex, first as a class, then as
# class probabilities
predict(class_st, tree_frogs_class_test)

#> # A tibble: 303 x 1
#>    .pred_class
#>    <fct>      
#>  1 low        
#>  2 mid        
#>  3 low        
#>  4 mid        
#>  5 low        
#>  6 low        
#>  7 mid        
#>  8 full       
#>  9 full       
#> 10 low        
#> # … with 293 more rows

predict(class_st, tree_frogs_class_test, type = "prob")

#> # A tibble: 303 x 3
#>    .pred_full   .pred_low .pred_mid
#>         <dbl>       <dbl>     <dbl>
#>  1   1.00e+ 0 0.0000608   0.000138 
#>  2   9.18e- 9 0.759       0.241    
#>  3   1.00e+ 0 0.00000243  0.0000429
#>  4   3.20e- 5 0.961       0.0386   
#>  5   1.00e+ 0 0.000000696 0.0000450
#>  6   1.00e+ 0 0.000000697 0.0000451
#>  7   5.09e-10 0.891       0.109    
#>  8   4.57e-13 0.308       0.692    
#>  9   9.49e-21 0.321       0.679    
#> 10   1.00e+ 0 0.000000778 0.0000603
#> # … with 293 more rows

# returning the member predictions as well
predict(
  class_st, 
  tree_frogs_class_test, 
  type = "prob", 
  members = TRUE
)

#> # A tibble: 303 x 36
#>    .pred_full .pred_low .pred_mid .pred_low_class… .pred_low_class…
#>         <dbl>     <dbl>     <dbl>            <dbl>            <dbl>
#>  1   1.00e+ 0   6.08e-5 0.000138             0.212           0.209 
#>  2   9.18e- 9   7.59e-1 0.241                0.415           0.515 
#>  3   1.00e+ 0   2.43e-6 0.0000429            0.212           0.0862
#>  4   3.20e- 5   9.61e-1 0.0386               0.458           0.562 
#>  5   1.00e+ 0   6.96e-7 0.0000450            0.212           0.0405
#>  6   1.00e+ 0   6.97e-7 0.0000451            0.212           0.0405
#>  7   5.09e-10   8.91e-1 0.109                0.409           0.500 
#>  8   4.57e-13   3.08e-1 0.692                0.416           0.510 
#>  9   9.49e-21   3.21e-1 0.679                0.276           0.277 
#> 10   1.00e+ 0   7.78e-7 0.0000603            0.212           0.0250
#> # … with 293 more rows, and 31 more variables: .pred_low_class_res_rf08 <dbl>,
#> #   .pred_low_class_res_rf05 <dbl>, .pred_low_class_res_rf07 <dbl>,
#> #   .pred_low_class_res_rf04 <dbl>, .pred_low_class_res_rf10 <dbl>,
#> #   .pred_low_class_res_rf03 <dbl>, .pred_low_class_res_rf02 <dbl>,
#> #   .pred_low_class_res_rf01 <dbl>, .pred_low_class_res_rf06 <dbl>,
#> #   .pred_mid_class_res_nn1 <dbl>, .pred_mid_class_res_rf09 <dbl>,
#> #   .pred_mid_class_res_rf08 <dbl>, .pred_mid_class_res_rf05 <dbl>,
#> #   .pred_mid_class_res_rf07 <dbl>, .pred_mid_class_res_rf04 <dbl>,
#> #   .pred_mid_class_res_rf10 <dbl>, .pred_mid_class_res_rf03 <dbl>,
#> #   .pred_mid_class_res_rf02 <dbl>, .pred_mid_class_res_rf01 <dbl>,
#> #   .pred_mid_class_res_rf06 <dbl>, .pred_full_class_res_nn1 <dbl>,
#> #   .pred_full_class_res_rf09 <dbl>, .pred_full_class_res_rf08 <dbl>,
#> #   .pred_full_class_res_rf05 <dbl>, .pred_full_class_res_rf07 <dbl>,
#> #   .pred_full_class_res_rf04 <dbl>, .pred_full_class_res_rf10 <dbl>,
#> #   .pred_full_class_res_rf03 <dbl>, .pred_full_class_res_rf02 <dbl>,
#> #   .pred_full_class_res_rf01 <dbl>, .pred_full_class_res_rf06 <dbl>

# }