Augment data with information from a model_stack object

Augment accepts a model object and a dataset and adds information about each observation in the dataset.

# S3 method for model_stack
augment(x, data = x[["train"]], ...)

Arguments

x
A model_stack object

data
A data.frame-like object containing all of the necessary predictors that the model stack was trained with.

...
Additional arguments. Currently ignored.

Value

A tibble::tibble() with additional information on each observation appended.

Example Data

These examples make use of data and resampling objects exported with the package. All of them are derived from test cases making use of the penguins dataset from Allison Horst's palmerpenguins package.

The penguins_train and penguins_test objects are subsets of the penguins data for using in training and testing, respectively.

Objects containing the substring _res_ are tune_results objects for model specifications. The reg prefix indicates that the model definition is for use in regression, class indicates multinomial classification, and log (as in logistic) indicates binomial classification. The suffix indicates the model definition type. All of these tune_results objects are fitted on a 5-fold cross validation of the penguins_train data.

For the regression setting, these tune_results objects reflect models specified to fit body_mass_g using all of the other variables as predictors. These objects include:

reg_res_lr: Fitted resamples for a linear regression model

reg_res_sp: Tuning results for a splines model

reg_res_svm: Tuning results for a support vector machine model

In the multinomial classification setting, the relevant objects reflect models specified to fit year (as a factor) using all of the other variables as predictors. These objects include:

class_res_nn: Fitted resamples for a neural network model

class_res_rf: Tuning results for a random forest model

In the binomial classification setting, the relevant objects reflect models specified to fit sex using all of the other variables as predictors. These objects include:

log_res_nn: Fitted resamples for a neural network model

log_res_rf: Tuning results for a random forest model

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

See also

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

# to use the `tidy` method, the appropriate glance
# methods must be loaded for the member model tidiers.
# in this example, they come from broom!
library(broom)

# build a regression model stack
st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()
  
# check the model stack out!
tidy(st)

#> # A tibble: 4 x 13
#>   member r.squared adj.r.squared sigma statistic    p.value    df logLik   AIC
#>   <chr>      <dbl>         <dbl> <dbl>     <dbl>      <dbl> <dbl>  <dbl> <dbl>
#> 1 reg_r…     0.872         0.867  298.     163.   1.14e-100    10 -1773. 3570.
#> 2 reg_r…     0.875         0.867  298.     109.   2.13e- 96    15 -1770. 3575.
#> 3 reg_r…     0.884         0.868  296.      53.8  3.59e- 85    31 -1760. 3587.
#> 4 ensem…    NA            NA       NA       NA   NA            NA    NA    NA 
#> # … with 4 more variables: BIC <dbl>, deviance <dbl>, df.residual <int>,
#> #   nobs <int>

glance(st)

#> # A tibble: 1 x 12
#>   r.squared adj.r.squared sigma statistic p.value    df logLik   AIC   BIC
#>       <dbl>         <dbl> <dbl>     <dbl>   <dbl> <int>  <dbl> <dbl> <dbl>
#> 1        NA            NA    NA        NA      NA    NA     NA    NA    NA
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>

augment(st)

#> # A tibble: 250 x 9
#>    species island bill_length_mm bill_depth_mm flipper_length_… body_mass_g
#>    <fct>   <fct>           <dbl>         <dbl>            <int>       <int>
#>  1 Adelie  Torge…           39.1          18.7              181        3750
#>  2 Adelie  Torge…           39.5          17.4              186        3800
#>  3 Adelie  Torge…           40.3          18                195        3250
#>  4 Adelie  Torge…           36.7          19.3              193        3450
#>  5 Adelie  Torge…           39.3          20.6              190        3650
#>  6 Adelie  Torge…           38.9          17.8              181        3625
#>  7 Adelie  Torge…           39.2          19.6              195        4675
#>  8 Adelie  Torge…           41.1          17.6              182        3200
#>  9 Adelie  Torge…           38.6          21.2              191        3800
#> 10 Adelie  Torge…           34.6          21.1              198        4400
#> # … with 240 more rows, and 3 more variables: sex <fct>, year <fct>,
#> #   .fitted$.pred <dbl>

# }