Evaluate a data stack

Evaluates a stack by performing regularization on the out-of-sample predictions to determine coefficients for the combining of predictions from ensemble members.

blend_predictions(data_stack, penalty = 10^(-6:-1), verbose = FALSE, ...)

Arguments

data_stack
A data_stack object

penalty
A numeric vector of proposed penalty values used in member weighting. Higher penalties will generally result in fewer members being included in the resulting model stack, and vice versa. This argument will be tuned on unless a single penalty value is given.

verbose
A logical for logging results as they are generated. Despite this argument, warnings and errors are always shown.

...
Additional arguments. Currently ignored.

Value

A model_stack object—while model_stacks largely contain the same elements as data_stacks, the primary data objects shift from the assessment set predictions to the member models.

Example Data

This package provides some resampling objects and datasets for use in examples and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal 7ish days if they detect potential predator threat. Researchers wanted to determine how, and when, these tree frog embryos were able to detect stimulus from their environment. To do so, they subjected the embryos at varying developmental stages to "predator stimulus" by jiggling the embryos with a blunt probe. Beforehand, though some of the embryos were treated with gentamicin, a compound that knocks out their lateral line (a sensory organ.) Researcher Julie Jung and her crew found that these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily a representative or unbiased subset of the complete dataset, and is only for demonstrative purposes.

reg_folds and class_folds are rset cross-fold validation objects from rsample, splitting the training data into for the regression and classification model objects, respectively. tree_frogs_reg_test and tree_frogs_class_test are the analogous testing sets.

reg_res_lr, reg_res_svm, and reg_res_sp contain regression tuning results for a linear regression, support vector machine, and spline model, respectively, fitting latency (i.e. how long the embryos took to hatch in response to the jiggle) in the tree_frogs data, using most all of the other variables as predictors. Note that the data underlying these models is filtered to include data only from embryos that hatched in response to the stimulus.

class_res_rf and class_res_nn contain multiclass classification tuning results for a random forest and neural network classification model, respectively, fitting reflex (a measure of ear function) in the data using most all of the other variables as predictors.

log_res_rf and log_res_nn, contain binary classification tuning results for a random forest and neural network classification model, respectively, fitting hatched (whether or not the embryos hatched in response to the stimulus) using most all of the other variables as predictors.

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

See also

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the objects used in these examples!

# put together a data stack
reg_st <- 
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_svm) %>%
  add_candidates(reg_res_sp)
  
reg_st

#> # A data stack with 3 model definitions and 15 candidate members:
#> #   reg_res_lr: 1 sub-model
#> #   reg_res_svm: 5 sub-models
#> #   reg_res_sp: 9 sub-models
#> # Outcome: latency (numeric)

# evaluate the data stack
reg_st %>%
  blend_predictions()

#> ── A stacked ensemble model ─────────────────────────────────────

#> 
#> Out of 15 possible blending coefficients, the ensemble used 5.
#> Lasso penalty: 0.1.

#> 
#> The 5 highest weighted members are:

#> # A tibble: 5 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_svm3 svm_rbf     0.987
#> 2 reg_res_svm4 svm_rbf     0.640
#> 3 reg_res_svm1 svm_rbf     0.405
#> 4 reg_res_sp9  linear_reg  0.294
#> 5 reg_res_svm5 svm_rbf     0.293

#> 
#> Members have not yet been fitted with `fit_members()`.

# include fewer models by proposing
# higher penalties
reg_st %>% blend_predictions(penalty = c(.5, 1))

#> ── A stacked ensemble model ─────────────────────────────────────

#> 
#> Out of 15 possible blending coefficients, the ensemble used 3.
#> Lasso penalty: 1.

#> 
#> The 3 highest weighted members are:

#> # A tibble: 3 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_svm3 svm_rbf     0.972
#> 2 reg_res_svm5 svm_rbf     0.337
#> 3 reg_res_sp9  linear_reg  0.284

#> 
#> Members have not yet been fitted with `fit_members()`.

  
# do the same with multinomial classification models
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions()
  
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> 
#> Out of 22 possible blending coefficients, the ensemble used 4.
#> Lasso penalty: 0.01.

#> Across the 3 classes, there are an average of 2 coefficients per class.

#> 
#> The 4 highest weighted member classes are:

#> # A tibble: 4 x 4
#>   member                    type        weight class
#>   <chr>                     <chr>        <dbl> <chr>
#> 1 .pred_full_class_res_nn1  mlp         20.0   full 
#> 2 .pred_mid_class_res_rf02  rand_forest  3.53  mid  
#> 3 .pred_full_class_res_rf01 rand_forest  1.52  full 
#> 4 .pred_full_class_res_rf09 rand_forest  0.259 full 

#> 
#> Members have not yet been fitted with `fit_members()`.

# ...or binomial classification models
log_st <-
  stacks() %>%
  add_candidates(log_res_nn) %>%
  add_candidates(log_res_rf) %>%
  blend_predictions()
  
log_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> 
#> Out of 11 possible blending coefficients, the ensemble used 3.
#> Lasso penalty: 0.1.

#> 
#> The 3 highest weighted member classes are:

#> # A tibble: 3 x 3
#>   member                 type        weight
#>   <chr>                  <chr>        <dbl>
#> 1 .pred_yes_log_res_nn1  mlp          5.08 
#> 2 .pred_yes_log_res_rf03 rand_forest  1.00 
#> 3 .pred_yes_log_res_rf01 rand_forest  0.287

#> 
#> Members have not yet been fitted with `fit_members()`.

# }