Model predictions

Apply a model stack to create different types of predictions.

# S3 method for model_stack
predict(object, new_data, type = NULL, members = FALSE, opts = list(), ...)

Arguments

object
A model stack with fitted members.

new_data
A rectangular data object, such as a data frame.

type
Format of returned predicted values—one of "numeric", "class", or "prob". When NULL, predict() will choose an appropriate value based on the model's mode.

members
Logical. Whether or not to additionally return the predictions for each of the ensemble members.

opts
A list of optional arguments to the underlying predict function passed on to parsnip::predict.model_fit for each member.

...
Additional arguments. Currently ignored.

Example Data

These examples make use of data and resampling objects exported with the package. All of them are derived from test cases making use of the penguins dataset from Allison Horst's palmerpenguins package.

The penguins_train and penguins_test objects are subsets of the penguins data for using in training and testing, respectively.

Objects containing the substring _res_ are tune_results objects for model specifications. The reg prefix indicates that the model definition is for use in regression, class indicates multinomial classification, and log (as in logistic) indicates binomial classification. The suffix indicates the model definition type. All of these tune_results objects are fitted on a 5-fold cross validation of the penguins_train data.

For the regression setting, these tune_results objects reflect models specified to fit body_mass_g using all of the other variables as predictors. These objects include:

reg_res_lr: Fitted resamples for a linear regression model

reg_res_sp: Tuning results for a splines model

reg_res_svm: Tuning results for a support vector machine model

In the multinomial classification setting, the relevant objects reflect models specified to fit year (as a factor) using all of the other variables as predictors. These objects include:

class_res_nn: Fitted resamples for a neural network model

class_res_rf: Tuning results for a random forest model

In the binomial classification setting, the relevant objects reflect models specified to fit sex using all of the other variables as predictors. These objects include:

log_res_nn: Fitted resamples for a neural network model

log_res_rf: Tuning results for a random forest model

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

data(penguins_test)

#> Warning: data set ‘penguins_test’ not found

# build and fit a regression model stack
reg_st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()

reg_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 11 possible candidates, the ensemble contains 3 members.

#> The highest weighted members are:

#> # A tibble: 3 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_sp10 linear_reg  0.379
#> 2 reg_res_lr1  linear_reg  0.335
#> 3 reg_res_sp06 linear_reg  0.209

# predict on the penguins testing data
predict(reg_st, penguins_test)

#> Error in rlang::call2(paste0("predict_members_", object[["mode"]]), model_stack = object,     coefs = coefs, new_data = new_data, opts = opts, type = member_type): object 'penguins_test' not found

# include the predictions from the members
predict(reg_st, penguins_test, members = TRUE)

#> Error in rlang::call2(paste0("predict_members_", object[["mode"]]), model_stack = object,     coefs = coefs, new_data = new_data, opts = opts, type = member_type): object 'penguins_test' not found

# build and fit a classification model stack
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()

#> ! Fold2: model 1/1: from glmnet Fortran code (error code -87); Convergence for 87...

#> ! Fold5: model 1/1: from glmnet Fortran code (error code -84); Convergence for 84...

 
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 22 possible candidates, the ensemble contains 5 members.

#> Across the 3 classes, there are an average of 2.5 members per class.

#> The highest weighted member classes are:

#> # A tibble: 5 x 4
#>   member                    type         weight class
#>   <chr>                     <chr>         <dbl> <chr>
#> 1 .pred_full_class_res_nn1  mlp         19.4    full 
#> 2 .pred_mid_class_res_rf06  rand_forest  3.47   mid  
#> 3 .pred_full_class_res_rf05 rand_forest  1.65   full 
#> 4 .pred_full_class_res_rf03 rand_forest  0.315  full 
#> 5 .pred_full_class_res_rf02 rand_forest  0.0477 full 

# predict year, first as a class, then as
# class probabilities
predict(class_st, penguins_test)

#> Error in rlang::call2(paste0("predict_members_", object[["mode"]]), model_stack = object,     coefs = coefs, new_data = new_data, opts = opts, type = member_type): object 'penguins_test' not found

predict(class_st, penguins_test, type = "prob")

#> Error in rlang::call2(paste0("predict_members_", object[["mode"]]), model_stack = object,     coefs = coefs, new_data = new_data, opts = opts, type = member_type): object 'penguins_test' not found

# returning the member predictions as well
predict(
  class_st, 
  penguins_test, 
  type = "prob", 
  members = TRUE
)

#> Error in rlang::call2(paste0("predict_members_", object[["mode"]]), model_stack = object,     coefs = coefs, new_data = new_data, opts = opts, type = member_type): object 'penguins_test' not found

# }