Model predictions

Apply a model stack to create different types of predictions.

# S3 method for model_stack
predict(object, new_data, type = NULL, members = FALSE, opts = list(), ...)

Arguments

object
A model stack with fitted members.

new_data
A rectangular data object, such as a data frame.

type
Format of returned predicted values—one of "numeric", "class", or "prob". When NULL, predict() will choose an appropriate value based on the model's mode.

members
Logical. Whether or not to additionally return the predictions for each of the ensemble members.

opts
A list of optional arguments to the underlying predict function passed on to parsnip::predict.model_fit for each member.

...
Additional arguments. Currently ignored.

Example Data

These examples make use of data and resampling objects exported with the package. All of them are derived from test cases making use of the penguins dataset from Allison Horst's palmerpenguins package.

The penguins_train and penguins_test objects are subsets of the penguins data for using in training and testing, respectively.

Objects containing the substring _res_ are tune_results objects for model specifications. The reg prefix indicates that the model definition is for use in regression, class indicates multinomial classification, and log (as in logistic) indicates binomial classification. The suffix indicates the model definition type. All of these tune_results objects are fitted on a 5-fold cross validation of the penguins_train data.

For the regression setting, these tune_results objects reflect models specified to fit body_mass_g using all of the other variables as predictors. These objects include:

reg_res_lr: Fitted resamples for a linear regression model

reg_res_sp: Tuning results for a splines model

reg_res_svm: Tuning results for a support vector machine model

In the multinomial classification setting, the relevant objects reflect models specified to fit year (as a factor) using all of the other variables as predictors. These objects include:

class_res_nn: Fitted resamples for a neural network model

class_res_rf: Tuning results for a random forest model

In the binomial classification setting, the relevant objects reflect models specified to fit sex using all of the other variables as predictors. These objects include:

log_res_nn: Fitted resamples for a neural network model

log_res_rf: Tuning results for a random forest model

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

data(penguins_test)

# build and fit a regression model stack
reg_st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()

reg_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 11 possible candidates, the ensemble contains 3 members.

#> The highest weighted members are:

#> # A tibble: 3 x 3
#>   member       type       weight
#>   <chr>        <chr>       <dbl>
#> 1 reg_res_lr1  linear_reg 0.641 
#> 2 reg_res_sp06 linear_reg 0.331 
#> 3 reg_res_sp08 linear_reg 0.0107

# predict on the penguins testing data
predict(reg_st, penguins_test)

#> # A tibble: 83 x 1
#>    .pred
#>    <dbl>
#>  1 4276.
#>  2 3396.
#>  3 3707.
#>  4 3877.
#>  5 3277.
#>  6 3839.
#>  7 3900.
#>  8 3568.
#>  9 4067.
#> 10 4204.
#> # … with 73 more rows

# include the predictions from the members
predict(reg_st, penguins_test, members = TRUE)

#> # A tibble: 83 x 4
#>    .pred reg_res_lr1 reg_res_sp08 reg_res_sp06
#>    <dbl>       <dbl>        <dbl>        <dbl>
#>  1 4276.       4265.        4216.        4293.
#>  2 3396.       3370.        3431.        3394.
#>  3 3707.       3698.        3699.        3690.
#>  4 3877.       3885.        3894.        3833.
#>  5 3277.       3299.        3324.        3175.
#>  6 3839.       3854.        3830.        3781.
#>  7 3900.       3934.        3902.        3810.
#>  8 3568.       3519.        3532.        3620.
#>  9 4067.       4081.        4100.        4025.
#> 10 4204.       4258.        4166.        4091.
#> # … with 73 more rows

# build and fit a classification model stack
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()

#> ! Fold1: model 1/1: from glmnet Fortran code (error code -58); Convergence for 58...

#> ! Fold2: model 1/1: from glmnet Fortran code (error code -65); Convergence for 65...

#> ! Fold3: model 1/1: from glmnet Fortran code (error code -68); Convergence for 68...

#> ! Fold4: model 1/1: from glmnet Fortran code (error code -62); Convergence for 62...

#> ! Fold5: model 1/1: from glmnet Fortran code (error code -63); Convergence for 63...

 
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> Out of 22 possible candidates, the ensemble contains 13 members.

#> Across the 3 classes, there are an average of 4.33 members per class.

#> The highest weighted member classes are:

#> # A tibble: 10 x 4
#>    member                    type        weight class
#>    <chr>                     <chr>        <dbl> <chr>
#>  1 .pred_2009_class_res_rf01 rand_forest  7.19  2007 
#>  2 .pred_2009_class_res_rf01 rand_forest  5.35  2008 
#>  3 .pred_2009_class_res_nn1  mlp          3.31  2008 
#>  4 .pred_2009_class_res_nn1  mlp          3.06  2009 
#>  5 .pred_2009_class_res_rf06 rand_forest  2.46  2009 
#>  6 .pred_2009_class_res_rf04 rand_forest  2.07  2009 
#>  7 .pred_2008_class_res_rf02 rand_forest  1.12  2008 
#>  8 .pred_2008_class_res_nn1  mlp          0.899 2008 
#>  9 .pred_2008_class_res_rf05 rand_forest  0.836 2009 
#> 10 .pred_2008_class_res_rf07 rand_forest  0.727 2008 

# predict year, first as a class, then as
# class probabilities
predict(class_st, penguins_test)

#> # A tibble: 83 x 1
#>    .pred_class
#>    <fct>      
#>  1 2009       
#>  2 2008       
#>  3 2007       
#>  4 2009       
#>  5 2007       
#>  6 2007       
#>  7 2007       
#>  8 2007       
#>  9 2008       
#> 10 2007       
#> # … with 73 more rows

predict(class_st, penguins_test, type = "prob")

#> # A tibble: 83 x 3
#>    .pred_2007 .pred_2008 .pred_2009
#>         <dbl>      <dbl>      <dbl>
#>  1      0.365      0.259      0.377
#>  2      0.290      0.383      0.327
#>  3      0.447      0.283      0.270
#>  4      0.314      0.309      0.377
#>  5      0.453      0.282      0.265
#>  6      0.413      0.321      0.266
#>  7      0.404      0.381      0.214
#>  8      0.426      0.296      0.278
#>  9      0.151      0.564      0.285
#> 10      0.383      0.309      0.307
#> # … with 73 more rows

# returning the member predictions as well
predict(
  class_st, 
  penguins_test, 
  type = "prob", 
  members = TRUE
)

#> # A tibble: 83 x 24
#>    .pred_2007 .pred_2008 .pred_2009 .pred_2007_clas… .pred_2007_clas…
#>         <dbl>      <dbl>      <dbl>            <dbl>            <dbl>
#>  1      0.365      0.259      0.377            0.319            0.574
#>  2      0.290      0.383      0.327            0.289            0.454
#>  3      0.447      0.283      0.270            0.404            0.354
#>  4      0.314      0.309      0.377            0.266            0.267
#>  5      0.453      0.282      0.265            0.483            0.411
#>  6      0.413      0.321      0.266            0.416            0.281
#>  7      0.404      0.381      0.214            0.316            0.437
#>  8      0.426      0.296      0.278            0.496            0.381
#>  9      0.151      0.564      0.285            0.264            0.218
#> 10      0.383      0.309      0.307            0.284            0.571
#> # … with 73 more rows, and 19 more variables: .pred_2007_class_res_rf02 <dbl>,
#> #   .pred_2007_class_res_rf07 <dbl>, .pred_2007_class_res_rf05 <dbl>,
#> #   .pred_2007_class_res_rf04 <dbl>, .pred_2007_class_res_rf06 <dbl>,
#> #   .pred_2008_class_res_rf01 <dbl>, .pred_2008_class_res_nn1 <dbl>,
#> #   .pred_2008_class_res_rf02 <dbl>, .pred_2008_class_res_rf07 <dbl>,
#> #   .pred_2008_class_res_rf05 <dbl>, .pred_2008_class_res_rf04 <dbl>,
#> #   .pred_2008_class_res_rf06 <dbl>, .pred_2009_class_res_rf01 <dbl>,
#> #   .pred_2009_class_res_nn1 <dbl>, .pred_2009_class_res_rf02 <dbl>,
#> #   .pred_2009_class_res_rf07 <dbl>, .pred_2009_class_res_rf05 <dbl>,
#> #   .pred_2009_class_res_rf04 <dbl>, .pred_2009_class_res_rf06 <dbl>

# }