Predicting with a model stack

Apply a model stack to create different types of predictions.

# S3 method for model_stack
predict(object, new_data, type = NULL, members = FALSE, opts = list(), ...)

Arguments

object
A model stack with fitted members.

new_data
A rectangular data object, such as a data frame.

type
Format of returned predicted values—one of "numeric", "class", or "prob". When NULL, predict() will choose an appropriate value based on the model's mode.

members
Logical. Whether or not to additionally return the predictions for each of the ensemble members.

opts
A list of optional arguments to the underlying predict function passed on to parsnip::predict.model_fit for each member.

...
Additional arguments. Currently ignored.

Example Data

This package provides some resampling objects and datasets for use in examples and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal 7ish days if they detect potential predator threat. Researchers wanted to determine how, and when, these tree frog embryos were able to detect stimulus from their environment. To do so, they subjected the embryos at varying developmental stages to "predator stimulus" by jiggling the embryos with a blunt probe. Beforehand, though some of the embryos were treated with gentamicin, a compound that knocks out their lateral line (a sensory organ.) Researcher Julie Jung and her crew found that these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily a representative or unbiased subset of the complete dataset, and is only for demonstrative purposes.

reg_folds and class_folds are rset cross-fold validation objects from rsample, splitting the training data into for the regression and classification model objects, respectively. tree_frogs_reg_test and tree_frogs_class_test are the analogous testing sets.

reg_res_lr, reg_res_svm, and reg_res_sp contain regression tuning results for a linear regression, support vector machine, and spline model, respectively, fitting latency (i.e. how long the embryos took to hatch in response to the jiggle) in the tree_frogs data, using most all of the other variables as predictors. Note that the data underlying these models is filtered to include data only from embryos that hatched in response to the stimulus.

class_res_rf and class_res_nn contain multiclass classification tuning results for a random forest and neural network classification model, respectively, fitting reflex (a measure of ear function) in the data using most all of the other variables as predictors.

log_res_rf and log_res_nn, contain binary classification tuning results for a random forest and neural network classification model, respectively, fitting hatched (whether or not the embryos hatched in response to the stimulus) using most all of the other variables as predictors.

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

data(tree_frogs_reg_test)
data(tree_frogs_class_test)

# build and fit a regression model stack
reg_st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()

reg_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> 
#> Out of 10 possible candidate members, the ensemble retained 3.
#> Lasso penalty: 0.01.

#> 
#> The 3 highest weighted members are:

#> # A tibble: 3 x 3
#>   member      type       weight
#>   <chr>       <chr>       <dbl>
#> 1 reg_res_lr1 linear_reg  0.355
#> 2 reg_res_sp9 linear_reg  0.304
#> 3 reg_res_sp5 linear_reg  0.261

# predict on the tree frogs testing data
predict(reg_st, tree_frogs_reg_test)

#> # A tibble: 143 x 1
#>    .pred
#>    <dbl>
#>  1  40.3
#>  2 111. 
#>  3  90.6
#>  4  33.7
#>  5  75.3
#>  6  90.0
#>  7 122. 
#>  8  82.4
#>  9  37.5
#> 10  77.3
#> # … with 133 more rows

# include the predictions from the members
predict(reg_st, tree_frogs_reg_test, members = TRUE)

#> # A tibble: 143 x 4
#>    .pred reg_res_lr1 reg_res_sp5 reg_res_sp9
#>    <dbl>       <dbl>       <dbl>       <dbl>
#>  1  40.3        38.0        34.3        36.9
#>  2 111.        124.        117.         98.6
#>  3  90.6        84.5        92.4        98.2
#>  4  33.7        35.3        28.0        23.8
#>  5  75.3        79.0        77.0        67.6
#>  6  90.0        83.7        92.2        97.3
#>  7 122.        118.        139.        123. 
#>  8  82.4        80.0        79.2        88.0
#>  9  37.5        36.5        30.7        32.7
#> 10  77.3        79.4        78.0        72.8
#> # … with 133 more rows

# build and fit a classification model stack
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()

#> ! Bootstrap14: model 1/1: from glmnet Fortran code (error code -84); Convergence for 84...

 
class_st

#> ── A stacked ensemble model ─────────────────────────────────────

#> 
#> Out of 22 possible candidate members, the ensemble retained 19.
#> Lasso penalty: 1e-04.

#> Across the 3 classes, there are an average of 6.33 coefficients per class.

#> 
#> The 10 highest weighted member classes are:

#> # A tibble: 10 x 4
#>    member                    type        weight class
#>    <chr>                     <chr>        <dbl> <chr>
#>  1 .pred_full_class_res_nn1  mlp          81.1  full 
#>  2 .pred_mid_class_res_rf04  rand_forest  33.4  mid  
#>  3 .pred_mid_class_res_rf05  rand_forest  29.5  low  
#>  4 .pred_full_class_res_nn1  mlp          18.0  mid  
#>  5 .pred_mid_class_res_rf09  rand_forest  16.9  low  
#>  6 .pred_full_class_res_rf04 rand_forest  16.2  mid  
#>  7 .pred_mid_class_res_rf10  rand_forest   6.32 low  
#>  8 .pred_mid_class_res_rf02  rand_forest   6.16 mid  
#>  9 .pred_mid_class_res_rf03  rand_forest   3.26 low  
#> 10 .pred_mid_class_res_rf08  rand_forest   2.28 mid  

# predict reflex, first as a class, then as
# class probabilities
predict(class_st, tree_frogs_class_test)

#> # A tibble: 303 x 1
#>    .pred_class
#>    <fct>      
#>  1 low        
#>  2 mid        
#>  3 low        
#>  4 mid        
#>  5 low        
#>  6 low        
#>  7 mid        
#>  8 full       
#>  9 full       
#> 10 low        
#> # … with 293 more rows

predict(class_st, tree_frogs_class_test, type = "prob")

#> # A tibble: 303 x 3
#>    .pred_full .pred_low .pred_mid
#>         <dbl>     <dbl>     <dbl>
#>  1   1.00e+ 0  1.02e-13 0.0000704
#>  2   1.35e- 6  7.35e- 1 0.265    
#>  3   1.00e+ 0  6.74e-15 0.0000876
#>  4   2.82e- 4  9.11e- 1 0.0892   
#>  5   1.00e+ 0  2.76e-15 0.0000796
#>  6   1.00e+ 0  2.77e-15 0.0000797
#>  7   1.42e- 7  8.47e- 1 0.153    
#>  8   1.17e- 8  3.53e- 1 0.647    
#>  9   5.14e-14  2.39e- 1 0.761    
#> 10   1.00e+ 0  3.13e-15 0.0000868
#> # … with 293 more rows

# returning the member predictions as well
predict(
  class_st, 
  tree_frogs_class_test, 
  type = "prob", 
  members = TRUE
)

#> # A tibble: 303 x 33
#>    .pred_full .pred_low .pred_mid .pred_low_class… .pred_low_class…
#>         <dbl>     <dbl>     <dbl>            <dbl>            <dbl>
#>  1   1.00e+ 0  1.02e-13 0.0000704           0.195          0.0782  
#>  2   1.35e- 6  7.35e- 1 0.265               0.519          0.608   
#>  3   1.00e+ 0  6.74e-15 0.0000876           0.0944         0.00960 
#>  4   2.82e- 4  9.11e- 1 0.0892              0.565          0.731   
#>  5   1.00e+ 0  2.76e-15 0.0000796           0.0445         0.000676
#>  6   1.00e+ 0  2.77e-15 0.0000797           0.0445         0.000676
#>  7   1.42e- 7  8.47e- 1 0.153               0.497          0.563   
#>  8   1.17e- 8  3.53e- 1 0.647               0.508          0.552   
#>  9   5.14e-14  2.39e- 1 0.761               0.257          0.227   
#> 10   1.00e+ 0  3.13e-15 0.0000868           0.0313         0.00312 
#> # … with 293 more rows, and 28 more variables: .pred_low_class_res_rf01 <dbl>,
#> #   .pred_low_class_res_rf03 <dbl>, .pred_low_class_res_rf10 <dbl>,
#> #   .pred_low_class_res_rf08 <dbl>, .pred_low_class_res_nn1 <dbl>,
#> #   .pred_low_class_res_rf04 <dbl>, .pred_low_class_res_rf07 <dbl>,
#> #   .pred_low_class_res_rf02 <dbl>, .pred_mid_class_res_rf05 <dbl>,
#> #   .pred_mid_class_res_rf09 <dbl>, .pred_mid_class_res_rf01 <dbl>,
#> #   .pred_mid_class_res_rf03 <dbl>, .pred_mid_class_res_rf10 <dbl>,
#> #   .pred_mid_class_res_rf08 <dbl>, .pred_mid_class_res_nn1 <dbl>,
#> #   .pred_mid_class_res_rf04 <dbl>, .pred_mid_class_res_rf07 <dbl>,
#> #   .pred_mid_class_res_rf02 <dbl>, .pred_full_class_res_rf05 <dbl>,
#> #   .pred_full_class_res_rf09 <dbl>, .pred_full_class_res_rf01 <dbl>,
#> #   .pred_full_class_res_rf03 <dbl>, .pred_full_class_res_rf10 <dbl>,
#> #   .pred_full_class_res_rf08 <dbl>, .pred_full_class_res_nn1 <dbl>,
#> #   .pred_full_class_res_rf04 <dbl>, .pred_full_class_res_rf07 <dbl>,
#> #   .pred_full_class_res_rf02 <dbl>

# }