Predicting with a model stack

Apply a model stack to create different types of predictions.

# S3 method for model_stack
predict(object, new_data, type = NULL, members = FALSE, opts = list(), ...)

Arguments

object
A model stack with fitted members outputted from fit_members().

new_data
A rectangular data object, such as a data frame.

type
Format of returned predicted values—one of "numeric", "class", or "prob". When NULL, predict() will choose an appropriate value based on the model's mode.

members
Logical. Whether or not to additionally return the predictions for each of the ensemble members.

opts
A list of optional arguments to the underlying predict function passed on to parsnip::predict.model_fit for each member.

...
Additional arguments. Currently ignored.

Example Data

This package provides some resampling objects and datasets for use in examples and vignettes derived from a study on 1212 red-eyed tree frog embryos!

Red-eyed tree frog (RETF) embryos can hatch earlier than their normal 7ish days if they detect potential predator threat. Researchers wanted to determine how, and when, these tree frog embryos were able to detect stimulus from their environment. To do so, they subjected the embryos at varying developmental stages to "predator stimulus" by jiggling the embryos with a blunt probe. Beforehand, though some of the embryos were treated with gentamicin, a compound that knocks out their lateral line (a sensory organ.) Researcher Julie Jung and her crew found that these factors inform whether an embryo hatches prematurely or not!

Note that the data included with the stacks package is not necessarily a representative or unbiased subset of the complete dataset, and is only for demonstrative purposes.

reg_folds and class_folds are rset cross-fold validation objects from rsample, splitting the training data into for the regression and classification model objects, respectively. tree_frogs_reg_test and tree_frogs_class_test are the analogous testing sets.

reg_res_lr, reg_res_svm, and reg_res_sp contain regression tuning results for a linear regression, support vector machine, and spline model, respectively, fitting latency (i.e. how long the embryos took to hatch in response to the jiggle) in the tree_frogs data, using most all of the other variables as predictors. Note that the data underlying these models is filtered to include data only from embryos that hatched in response to the stimulus.

class_res_rf and class_res_nn contain multiclass classification tuning results for a random forest and neural network classification model, respectively, fitting reflex (a measure of ear function) in the data using most all of the other variables as predictors.

log_res_rf and log_res_nn, contain binary classification tuning results for a random forest and neural network classification model, respectively, fitting hatched (whether or not the embryos hatched in response to the stimulus) using most all of the other variables as predictors.

See ?example_data to learn more about these objects, as well as browse the source code that generated them.

Examples

# \donttest{
# see the "Example Data" section above for
# clarification on the data and tuning results
# objects used in these examples!

data(tree_frogs_reg_test)
data(tree_frogs_class_test)

# build and fit a regression model stack
reg_st <-
  stacks() %>%
  add_candidates(reg_res_lr) %>%
  add_candidates(reg_res_sp) %>%
  blend_predictions() %>%
  fit_members()

reg_st
#> ── A stacked ensemble model ─────────────────────────────────────
#> 
#> Out of 10 possible candidate members, the ensemble retained 3.
#> Penalty: 0.1.
#> Mixture: 1.
#> 
#> The 3 highest weighted members are:
#> # A tibble: 3 x 3
#>   member         type       weight
#>   <chr>          <chr>       <dbl>
#> 1 reg_res_lr_1_1 linear_reg  0.349
#> 2 reg_res_sp_2_1 linear_reg  0.303
#> 3 reg_res_sp_8_1 linear_reg  0.267

# predict on the tree frogs testing data
predict(reg_st, tree_frogs_reg_test)
#> # A tibble: 143 x 1
#>    .pred
#>    <dbl>
#>  1  40.4
#>  2 111. 
#>  3  90.6
#>  4  33.8
#>  5  75.3
#>  6  90.0
#>  7 122. 
#>  8  82.4
#>  9  37.6
#> 10  77.3
#> # … with 133 more rows

# include the predictions from the members
predict(reg_st, tree_frogs_reg_test, members = TRUE)
#> # A tibble: 143 x 4
#>    .pred reg_res_lr_1_1 reg_res_sp_8_1 reg_res_sp_2_1
#>    <dbl>          <dbl>          <dbl>          <dbl>
#>  1  40.4           38.0           34.3           36.9
#>  2 111.           124.           117.            98.6
#>  3  90.6           84.5           92.4           98.2
#>  4  33.8           35.3           28.0           23.8
#>  5  75.3           79.0           77.0           67.6
#>  6  90.0           83.7           92.2           97.3
#>  7 122.           118.           139.           123. 
#>  8  82.4           80.0           79.2           88.0
#>  9  37.6           36.5           30.7           32.7
#> 10  77.3           79.4           78.0           72.8
#> # … with 133 more rows

# build and fit a classification model stack
class_st <-
  stacks() %>%
  add_candidates(class_res_nn) %>%
  add_candidates(class_res_rf) %>%
  blend_predictions() %>%
  fit_members()
#> ! Bootstrap05: preprocessor 1/1, model 1/1: from glmnet Fortran code (error code -61); ...
#> ! Bootstrap09: preprocessor 1/1, model 1/1: from glmnet Fortran code (error code -73); ...
#> ! Bootstrap11: preprocessor 1/1, model 1/1: from glmnet Fortran code (error code -65); ...
#> ! Bootstrap18: preprocessor 1/1, model 1/1: from glmnet Fortran code (error code -79); ...
#> ! Bootstrap19: preprocessor 1/1, model 1/1: from glmnet Fortran code (error code -77); ...
#> ! Bootstrap23: preprocessor 1/1, model 1/1: from glmnet Fortran code (error code -82); ...
#> Error in {    asNamespace("stacks")$fit_member(name = mem, wflows = model_stack[["model_defs"]],         members_map = members_map, train_dat = dat)} task 2 failed - "This engine requires some package installs: 'ranger'"
 
class_st
#> Error in eval(expr, envir, enclos) object 'class_st' not found

# predict reflex, first as a class, then as
# class probabilities
predict(class_st, tree_frogs_class_test)
#> Error in predict(class_st, tree_frogs_class_test) object 'class_st' not found
predict(class_st, tree_frogs_class_test, type = "prob")
#> Error in predict(class_st, tree_frogs_class_test, type = "prob") object 'class_st' not found

# returning the member predictions as well
predict(
  class_st, 
  tree_frogs_class_test, 
  type = "prob", 
  members = TRUE
)
#> Error in predict(class_st, tree_frogs_class_test, type = "prob", members = TRUE) object 'class_st' not found
# }